chore: codebase cleanup, renamed folders to app and api
