+++ MVP/Team#8/api/requirements.txt
@@ -1,0 +1,6 @@
+fastapi
+uvicorn[standard]
+torch
+numpy<2.0
+pydantic
+pandas
