+++ MVP/Team#8/app/entrypoint.sh
