+++ MVP/Team#8/app/playgrounds/urls.py
@@ -1,0 +1,5 @@
+from django.urls import path
+
+from .views import linear_regression
+
+urlpatterns = [path("linear_regression/", linear_regression, name="linear-regression")]
