--- MVP/Team#8/deno/tailwind.config.js
@@ -1,51 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-
-
-module.exports = {
-  mode: 'jit',
-  content: ["/templates/**/*.{html,js}"],
-  theme: {
-    extend: {},
-  },
-  plugins: [require("@tailwindcss/typography"),
-            require("daisyui")],
-
-  daisyui: {
-    themes: [
-      "light",
-      "dark",
-      "cupcake",
-      "bumblebee",
-      "emerald",
-      "corporate",
-      "synthwave",
-      "retro",
-      "cyberpunk",
-      "valentine",
-      "halloween",
-      "garden",
-      "forest",
-      "aqua",
-      "lofi",
-      "pastel",
-      "fantasy",
-      "wireframe",
-      "black",
-      "luxury",
-      "dracula",
-      "cmyk",
-      "autumn",
-      "business",
-      "acid",
-      "lemonade",
-      "night",
-      "coffee",
-      "winter",
-      "dim",
-      "nord",
-      "sunset",
-    ]},
-  
-  darkMode: 'selector',
-}
-
