--- MVP/Team#8/fastapi/requirements.txt
@@ -1,6 +1,0 @@
-fastapi
-uvicorn[standard]
-torch
-numpy
-pydantic
-pandas
