--- MVP/Team#8/redis.conf
@@ -1,1 +1,0 @@
-loglevel verbose
